==> code/0.wind_data_prep.Rmd <==
---
title: "0.daily_wind_data_prep"
author: "Dennis Kim"
date: '2022-06-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Preamble 

call relevant libraries 
```{r import_libraries}
# data wrangling 
library(dplyr)
library(readr)
library(data.table)
library(ezknitr)
library(here)
library(DT)
library(lubridate)
library(stringr)
library(bit64)
library(tidyr)
library(lubridate)
library(adehabitatLT)

# visualization
library(moveVis)
library(move)
library(ggplot2)
library(gganimate)
library(magick)
library(ggOceanMaps)
library(ggspatial)
library(grid)
library(transformr)
library(gifski)
library(magick)

# environmental layers 
library(raster)
library(rWind)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
```

# Environmental data

Import the raw data from your designated folder. 

## ocean data 
```{r ocean data}
# call the sf object of world map from rnaturalearth pacakge
world <- ne_countries(scale = "medium", returnclass = "sf")
class(world)

# view the Atlantic ocean world map 
ggplot(data = world) +
  geom_sf() +
  coord_sf(xlim = c(-80, 30), ylim = c(-70, 70), expand = FALSE) +
  theme_bw()  
```


## wind data
```{r wind data}
## set the time series 11 hrs interval
#dt <- seq(ymd_hms(paste(2017, 8, 8, 00, 00, 00, sep ="-")), ymd_hms(paste(2017, 11, 17, 23, 59, 59, sep = "-")), by = "11 hours")

# download Atlantic ocean dataset of a specified date range from GFS 
#wind_data <- wind.dl_2(dt, -80, 30, -70, 70)

# tidy the wind data as a data frame
#wind_data2 <- rWind::tidy(wind_data)

# save the wind_data 
#write_rds(wind_data2, path = here("data", "fall_wind.Rdata"))

# note: the above codes are the ones that make the fall_wind.Rdata - re-running the above codes take lots of time so I saved it Rdata for further convenient use

# call the saved wind_data rds
wind_data <- read_rds(here("data", "fall_wind.Rdata"))

# look at the summary of the data
summary(wind_data)
```
The wind data has velocity within 10m grid map. 


# Tidying the wind data frame
The current wind data is a lot (13848523 rows!)! We would group them by day
```{r monthly wind}
# first create a new column called month & day
wind = wind_data %>% mutate(month = month(time)) %>% mutate(day = day(time))

# convert the time as Date class 
wind$time <- as.Date(wind$time)

# create a data frame focus on month wind 
wind.time <- wind %>% group_by(lon, lat, time) %>% #group by lon, lat, time 
  summarise(u = median(ugrd10m, na.rm = TRUE), # calculate the median of eastward velocity per day 
            v = median(vgrd10m, na.rm = TRUE), # calculate the median of northward velocity per day
            velocity = median(speed, na.rm = TRUE) # calculate the median of speed per day 
            )

# convert the it to data frame
wind.time <- data.frame(wind.time)

# check
summary(wind.time)

# save the wind_data 
#write_rds(wind.time, path = here("data", "fall_wind_daily.Rdata"))

# since I already save the above code - here I recall the saved data so that I don't need to re-run the entire above code
wind.time.df <- read_rds(here("data", "fall_wind_daily.Rdata"))
```

# Wind visualization

This vector field has two problems: the data is at too high a resolution to read and the segments do not have arrows indicating the direction of the flow. 

To reduce the resolution of the data, we'll define a function *every_n()* that keeps one out of every *n* values in the data and drops the rest.
```{r reduce lines}
# create a function - keep 1 out of every 'by' values in vector x 
every_n <- function(x, by = 2){
  x <- sort(x)
  x[seq(1, length(x), by = by)]
}

# keep 1 of every 2 values in x and y in each vector map
keepx <- every_n(unique(wind.time$lon), by = 2)
keepy <- every_n(unique(wind.time$lat), by = 2)

# keep only those rows where x value is in keepx and y value is in keepy 
wind.time.df <- filter(wind.time, lon %in% keepx & lat %in% keepy)

# save the modified wind_data 
#write_rds(wind.time.df, path = here("data", "modified_fall_wind_daily.Rdata"))
```

## August visualization
```{r only august}
# visualize only august month to see if it works 
ggplot()+
  geom_segment(data = wind.time.df %>% filter(time == "2017-08-09"),
               aes(x = lon, xend = lon+u/10, y = lat, yend = lat+v/10), arrow = arrow(length = unit(0.1, "cm")))+
  geom_sf(data = world, fill = "grey85", col = 1)+
  coord_sf(xlim = c(-80, 30), ylim = c(-70, 70), expand = FALSE)+
  scale_x_continuous(breaks = c(-80, 30)) +
  theme_bw()+
  theme(axis.text = element_text(size = 11, colour = 1))+
  labs(x = NULL, y = NULL)
```

## daily visualization
```{r wind visualization fig.align="center", echo = FALSE,fig.width = 20, fig.height=20}
# visualization 
wind.vector = ggplot()+
  geom_raster(data = wind.time.df, aes(x = lon, y = lat, fill = velocity))+
  geom_segment(data = wind.time.df,
               aes(x = lon, xend = lon+u/10, y = lat, yend = lat+v/10), arrow = arrow(length = unit(0.1, "cm")))+
  scale_fill_gradientn(colours = cmocean('dense')(120), limits = c(0,12), 
                       na.value = "white", name = "Speed\n (m/s)")+
  geom_sf(data = world, fill = "grey85", col = 1)+
  coord_sf(xlim = c(-80, 30), ylim = c(-70, 70), expand = FALSE)+
  labs(x = NULL, y = NULL, title = "Date: {frame_time}")+
  transition_time(time) +
  ease_aes("linear")

# animate the visualization
animate(wind.vector, height = 1000, width = 1000, fps = 5)

# save the animation 
anim_save("fall_wind_daily.gif", width = 1000, height = 1000)
```

# Document Footer
```{r}
sessionInfo()
```
